+++ vgo/src/main/kotlin/com/jzbrooks/vgo/iv/ImageVectorGraphic.kt
@@ -1,4 +1,4 @@
-package com.jzbrooks.vgo.composable
+package com.jzbrooks.vgo.iv
 
 import com.jzbrooks.vgo.core.graphic.Element
 import com.jzbrooks.vgo.core.graphic.ElementVisitor
