+++ vgo/src/main/kotlin/com/jzbrooks/vgo/iv/ImageVectorGraphicWriter.kt
@@ -1,4 +1,4 @@
-package com.jzbrooks.vgo.composable
+package com.jzbrooks.vgo.iv
 
 import com.jzbrooks.vgo.core.Writer
 import com.jzbrooks.vgo.core.graphic.ClipPath
