+++ vgo/src/main/kotlin/com/jzbrooks/vgo/vd/SvgConverter.kt
@@ -1,6 +1,6 @@
 package com.jzbrooks.vgo.vd
 
-import com.jzbrooks.vgo.composable.ImageVectorGraphic
+import com.jzbrooks.vgo.iv.ImageVectorGraphic
 import com.jzbrooks.vgo.core.graphic.ContainerElement
 import com.jzbrooks.vgo.core.graphic.Element
 import com.jzbrooks.vgo.svg.ScalableVectorGraphic
